refactoring unit tests for future additions
hopefully fixed some linting issues
made unit tests runnable individually
separated out a utils file for testing db stuff
