+++ tests/persistence/test_db_pumped_milk.py
@@ -1,36 +1,9 @@
 import pytest
-import mariadb
+from pytest_mock import mocker
 
 from datetime import datetime
+from tests.persistence.utils import mock_init, mock_connection
 from boobjuice.persistence.database import PumpedMilk, DataAccessError, IllegalArgumentError
-
-class Connection:
-	def __init__(self, error=False, iter_list=None):
-		self.error = error
-		self.iter_list = iter_list
-
-	def cursor(self):
-		if self.error:
-			raise mariadb.Error
-		return Cursor(self.iter_list)
-	
-	def close(self):
-		pass
-
-class Cursor:
-	def __init__(self, iter_list=None):
-		self.iter_list = iter_list
-
-	def execute(self, query, params=None):
-		print(query)
-		marker_count = query.count('?')
-		if params is None:
-			assert marker_count == 0
-		else:
-			assert marker_count == len(params)
-	
-	def __iter__(self):
-		yield self.iter_list
 
 def test_init_db_error(mocker):
 	mock_connection(mocker, error=True)
@@ -44,24 +17,30 @@ def test_init_db_error(mocker):
 def test_init_success(mocker):
 	mock_connection(mocker)
 
-	global pumpedMilk
 	try:
 		pumpedMilk = PumpedMilk()
 	except DataAccessError:
 		pytest.fail('failed to init database', pytrace=True)
 
 def test_get_db_error(mocker):
+	mock_init(mocker)
 	mock_connection(mocker, error=True)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.get()
 		pytest.fail('should have failed to get from database', pytrace=True)
-	except DataAccessError:
-		assert True
+	except DataAccessError as e:
+		assert 'selecting' in e.message
 
 def test_get_success(mocker):
 	pumped_tuple = pumped_milk_tuple('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker, iter_list=pumped_tuple)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumped_list = pumpedMilk.get()
@@ -76,7 +55,11 @@ def test_get_success(mocker):
 
 def test_insert_missing_mass(mocker):
 	pumped_dict = pumped_milk_dict(None, None, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.insert(pumped_dict)
@@ -86,7 +69,11 @@ def test_insert_missing_mass(mocker):
 
 def test_insert_missing_duration(mocker):
 	pumped_dict = pumped_milk_dict(None, 200, None)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.insert(pumped_dict)
@@ -96,17 +83,25 @@ def test_insert_missing_duration(mocker):
 
 def test_insert_db_error(mocker):
 	pumped_dict = pumped_milk_dict(None, 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker, error=True)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.insert(pumped_dict)
 		pytest.fail('should have failed to insert to database', pytrace=True)
-	except DataAccessError:
-		assert True
+	except DataAccessError as e:
+		assert 'inserting' in e.message
 
 def test_insert_success_no_timestamp(mocker):
 	pumped_dict = pumped_milk_dict(None, 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.insert(pumped_dict)
@@ -116,7 +111,11 @@ def test_insert_success_no_timestamp(mocker):
 
 def test_insert_success_with_timestamp(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.insert(pumped_dict)
@@ -126,7 +125,11 @@ def test_insert_success_with_timestamp(mocker):
 
 def test_update_missing_timestamp(mocker):
 	pumped_dict = pumped_milk_dict(None, 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.update(pumped_dict)
@@ -136,7 +139,11 @@ def test_update_missing_timestamp(mocker):
 
 def test_update_missing_mass(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', None, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.update(pumped_dict)
@@ -146,7 +153,11 @@ def test_update_missing_mass(mocker):
 
 def test_update_missing_duration(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, None)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.update(pumped_dict)
@@ -156,17 +167,25 @@ def test_update_missing_duration(mocker):
 
 def test_update_db_error(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker, error=True)
 
+	pumpedMilk = PumpedMilk()
+
 	try:
-		pumpedMilk.insert(pumped_dict)
+		pumpedMilk.update(pumped_dict)
 		pytest.fail('should have failed to update database', pytrace=True)
-	except DataAccessError:
-		assert True
+	except DataAccessError as e:
+		assert 'updating' in e.message
 
 def test_update_success(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.update(pumped_dict)
@@ -176,7 +195,11 @@ def test_update_success(mocker):
 
 def test_delete_missing_timestamp(mocker):
 	pumped_dict = pumped_milk_dict(None, 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.delete(pumped_dict)
@@ -186,17 +209,25 @@ def test_delete_missing_timestamp(mocker):
 
 def test_delete_db_error(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker, error=True)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.delete(pumped_dict)
 		pytest.fail('should have failed to delete from database', pytrace=True)
-	except DataAccessError:
-		assert True
+	except DataAccessError as e:
+		assert 'deleting' in e.message
 
 def test_delete_success(mocker):
 	pumped_dict = pumped_milk_dict('2024-07-25T19:16', 200, 20)
+
+	mock_init(mocker)
 	mock_connection(mocker)
+
+	pumpedMilk = PumpedMilk()
 
 	try:
 		pumpedMilk.delete(pumped_dict)
@@ -204,39 +235,57 @@ def test_delete_success(mocker):
 	except DataAccessError:
 		pytest.fail('failed to delete from database', pytrace=True)
 
-def test_get_timestamp_optional_not_in():
+def test_get_timestamp_optional_not_in(mocker):
+	mock_init(mocker)
+
+	pumpedMilk = PumpedMilk()
 	assert pumpedMilk.get_timestamp({}, optional=True) is None
 
-def test_get_timestamp_optional_none():
+def test_get_timestamp_optional_none(mocker):
+	mock_init(mocker)
+	
+	pumpedMilk = PumpedMilk()
 	assert pumpedMilk.get_timestamp({'timestamp':None}, optional=True) is None
 
-def test_get_timestamp_required_not_in():
+def test_get_timestamp_required_not_in(mocker):
+	mock_init(mocker)
+	
+	pumpedMilk = PumpedMilk()
+
 	try:
 		pumpedMilk.get_timestamp({})
 	except IllegalArgumentError as e:
 		assert 'required' in e.message
 
-def test_get_timestamp_required_none():
+def test_get_timestamp_required_none(mocker):
+	mock_init(mocker)
+	
+	pumpedMilk = PumpedMilk()
+
 	try:
 		pumpedMilk.get_timestamp({'timestamp':None})
 	except IllegalArgumentError as e:
 		assert 'required' in e.message
 
-def test_get_timestamp_wrong_format():
+def test_get_timestamp_wrong_format(mocker):
+	mock_init(mocker)
+	
+	pumpedMilk = PumpedMilk()
+
 	try:
 		pumpedMilk.get_timestamp({'timestamp':'2024/07/25 19:16'})
 	except IllegalArgumentError as e:
 		assert 'invalid' in e.message
 
-def test_get_timestamp_success():
+def test_get_timestamp_success(mocker):
+	mock_init(mocker)
+	
+	pumpedMilk = PumpedMilk()
 	timestamp = pumpedMilk.get_timestamp({'timestamp':'2024-07-25T19:16'})
 	assert timestamp is not None
 
-def mock_connection(mocker, error=False, iter_list=None):
-	mocker.patch('boobjuice.persistence.database.get_connection', return_value=Connection(error=error, iter_list=iter_list))
-
 def pumped_milk_tuple(timestamp, mass, duration):
-	return (datetime.strptime(timestamp, pumpedMilk.ISO_8601), mass, duration)
+	return (datetime.strptime(timestamp, PumpedMilk.ISO_8601), mass, duration)
 
 def pumped_milk_dict(timestamp, mass, duration):
 	return {'timestamp':timestamp, 'mass':mass, 'duration':duration}
